skills/community/clawhub/c/compliance-checker-light/pyproject.toml
[project]
name = "compliance-checker"
version = "1.1.6"
description = "项目手续合规审查 Python 库"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "evob", email = "2502366203@qq.com"}
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "PyMuPDF>=1.23.0",
    "python-docx>=0.8.11",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",
    "python-dotenv>=0.21.0",
    "python-dateutil>=2.8.0",
    "numpy>=1.25.2,<2.6.0",
]

[project.urls]
Homepage = "https://github.com/evob-z/compliance_checker_light"
Repository = "https://github.com/evob-z/compliance_checker_light"
Issues = "https://github.com/evob-z/compliance_checker_light/issues"

[project.optional-dependencies]
# 本地 OCR（PaddleOCR，体积大但无需网络，可选安装）
# 注意：Paddle 3.3.0 在 Windows 上有 oneDNN 兼容性问题，建议使用 3.2.2
local-ocr = [
    "paddleocr>=3.0.0",
    "paddlepaddle>=3.0.0,<3.3.0",
]

# 云端 OCR（阿里云，轻量但需网络和付费，可选安装）
# alibabacloud_ocr_api20210707 3.x 为当前 OpenAPI 2021-07-07 对应实现（RecognizeGeneral 等）
cloud-ocr = [
    "alibabacloud_ocr_api20210707>=3.0.0",
    "alibabacloud-tea-openapi>=0.3.0",
]

# 完整安装（包含所有可选依赖）
all = [
    "paddleocr>=3.0.0",
    "paddlepaddle>=3.0.0,<3.3.0",
    "alibabacloud_ocr_api20210707>=3.0.0",
    "alibabacloud-tea-openapi>=0.3.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
]



[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = {find = {where = ["src"]}}

[tool.black]
line-length = 100
target-version = ['py310']

[tool.coverage.run]
source = ["src"]
omit = [
    # 初始化文件
    "*/__init__.py",
    # 配置文件
    "*/config/*.py",
    "*/config/settings.py",
    # 异常定义
    "*/core/exceptions.py",
    # YAML 兼容层（简单适配代码）
    "*/core/yaml_compat.py",
    # CLI 入口（集成测试覆盖）
    "*/cli.py",
    # 验证器（配置校验）
    "*/validators/*.py",
    # 报告生成（模板渲染）
    "*/report/*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "pass",
    "^\\s*\\.\\.\\.$",  # 省略号
]
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
skip_covered = false
